requirements.txt
psycopg[binary]
requests
python-dotenv
prefect
pytest
pytest-cov
supabase